fix: token expiry logic
