+++ apps/exceptions.py
@@ -14,8 +14,8 @@ from xerosdk.exceptions import (
 
 from apps.fyle.models import ExpenseGroup
 from apps.mappings.models import GeneralMapping, TenantMapping
+from apps.workspaces.helpers import invalidate_token
 from apps.workspaces.models import FyleCredential, Workspace, WorkspaceGeneralSettings, WorkspaceSchedule, XeroCredentials
-from apps.workspaces.actions import patch_integration_settings
 
 logger = logging.getLogger(__name__)
 logger.level = logging.INFO
@@ -50,6 +50,7 @@ def handle_view_exceptions():
                     kwargs["workspace_id"],
                     {"error": exception.response},
                 )
+                invalidate_token(kwargs["workspace_id"])
                 return Response(
                     data={"message": "Xero token expired workspace_id"},
                     status=status.HTTP_400_BAD_REQUEST,
@@ -73,7 +74,7 @@ def handle_view_exceptions():
                 InvalidClientError,
             ) as exception:
                 logger.info(exception)
-                patch_integration_settings(kwargs["workspace_id"], is_token_expired=True)
+                invalidate_token(kwargs["workspace_id"])
                 return Response(
                     data={"message": "Xero connection expired"},
                     status=status.HTTP_400_BAD_REQUEST,
