+++ apps/xero/queue.py
@@ -3,6 +3,7 @@ from typing import List
 import logging
 
 from django.db.models import Q
+from apps.workspaces.helpers import invalidate_token
 from django_q.models import Schedule
 from django_q.tasks import Chain
 from xerosdk.exceptions import InvalidGrant, UnsuccessfulAuthentication
@@ -181,6 +182,7 @@ def schedule_bills_creation(workspace_id: int, expense_group_ids: List[str], is_
                 xero_connection = XeroConnector(xero_credentials, workspace_id)
                 __create_chain_and_run(fyle_credentials, xero_connection, in_progress_expenses, workspace_id, chain_tasks, fund_source)
             except (UnsuccessfulAuthentication, InvalidGrant, XeroCredentials.DoesNotExist):
+                invalidate_token(workspace_id)
                 xero_connection = None
                 for task in chain_tasks:
                     task_log = TaskLog.objects.get(id=task['task_log_id'])
@@ -258,6 +260,7 @@ def schedule_bank_transaction_creation(
                 xero_connection = XeroConnector(xero_credentials, workspace_id)
                 __create_chain_and_run(fyle_credentials, xero_connection, in_progress_expenses, workspace_id, chain_tasks, fund_source)
             except (UnsuccessfulAuthentication, InvalidGrant, XeroCredentials.DoesNotExist):
+                invalidate_token(workspace_id)
                 xero_connection = None
                 for task in chain_tasks:
                     task_log = TaskLog.objects.get(id=task['task_log_id'])
